Move GW functionality into core's init file
